pyproject.toml
[project]
name = "ai-tutor-screener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.136.0",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn[standard]>=0.44.0",
]
